+++ core/src/main/java/YOGOSec/core/gui/InputListener.java
@@ -1,6 +1,6 @@
 package YOGOSec.core.gui;
 
-import YOGOSec.core.util.Vector2i;
+import YOGOSec.core.util.Vector2f;
 
 /**
  * @author Aritz Lopez
@@ -12,13 +12,13 @@ public interface InputListener {
 
     public boolean keyUp(int keycode);
 
-    public boolean touchDown(Vector2i point, int pointer, int button);
+    public boolean touchDown(Vector2f point, int pointer, int button);
 
-    public boolean touchUp(Vector2i point, int pointer, int button);
+    public boolean touchUp(Vector2f point, int pointer, int button);
 
-    public boolean touchDragged(Vector2i point, int pointer);
+    public boolean touchDragged(Vector2f point, int pointer);
 
-    public boolean mouseMoved(Vector2i point);
+    public boolean mouseMoved(Vector2f point);
 
     public boolean scrolled(int amount);
 }
